trade_journal_backend/requirements.txt
acres==0.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
distro==1.9.0
Django==5.0
django-allauth==65.3.0
django-auto-logout==0.5.1
django-axes==7.0.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-phonenumber-field==8.0.0
django-recaptcha==4.0.0
django-twilio==0.14.3.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environs==11.2.1
etelemetry==0.3.1
filelock==3.16.1
fitz==0.0.1.dev2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.0
idna==3.10
importlib_resources==6.4.5
isodate==0.6.1
jiter==0.8.0
looseversion==1.3.0
lxml==5.3.0
marshmallow==3.23.1
multidict==6.1.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.1
numpy==2.1.3
openai==1.56.0
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
phonenumbers==8.13.51
propcache==0.2.1
prov==2.0.1
psycopg2-binary==2.9.10
puremagic==1.28
pydantic==2.10.2
pydantic_core==2.27.1
pydot==3.0.3
PyJWT==2.10.1
PyMuPDF==1.24.14
pyparsing==3.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2024.2
pyxnat==1.6.2
rdflib==6.3.2
requests==2.32.3
scipy==1.14.1
sendgrid==6.11.0
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
starkbank-ecdsa==2.2.0
tqdm==4.67.1
traits==6.4.3
twilio==9.3.7
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.18.3
